add global sktypeface
